requirements.txt
asgiref==3.6.0
beautifulsoup4==4.12.2
bs4==0.0.1
Django==4.2.1
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
pytest==7.3.1
pytest-django==4.5.2
pytest-factoryboy==2.5.1
factory-boy==3.2.1
Faker==18.9.0
greenlet==2.0.2
marshmallow==3.19.0
packaging==23.1
Pillow==9.5.0
psycopg2-binary==2.9.6
PyJWT==2.7.0
python-dateutil==2.8.2
six==1.16.0
soupsieve==2.4.1
SQLAlchemy==1.4.31
sqlparse==0.4.4
